Backend/transGPT/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
blinker==1.8.2
boto3==1.34.136
botocore==1.34.136
braceexpand==0.1.7
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cpm-kernels==1.0.11
datasets==2.20.0
deepspeed==0.14.4
dill==0.3.8
einops==0.8.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.5.0
hjson==3.1.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
SwissArmyTransformer==0.4.11
sympy==1.12.1
tensorboardX==2.6.2.2
tokenizers==0.13.3
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.27.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
webdataset==0.2.86
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
